polish nvic set vectorfeature for nordic

+++ hal/targets/hal/TARGET_NORDIC/TARGET_NRF5/TARGET_MCU_NRF52832/analogin_api.c
@@ -47,7 +47,7 @@ void analogin_init(analogin_t *obj, PinName pin)
     ret_code_t ret_code;
     
 #ifndef HARDWIRE_ADC_INTERRUPT
-    NVIC_SetVector(SAADC_IRQn, SAADC_IRQHandler);
+    NVIC_SetVector(SAADC_IRQn, (uint32_t)SAADC_IRQHandler);
 #endif
     
     ret_code = nrf_drv_saadc_init(&saadc_config, analog_in_event_handler);
